+++ src/generator/index.js
@@ -24,6 +24,7 @@ export class Generator extends UIProxy {
   }
 
   findSystems(dir) {
+    log.debug('court.judge(\'%s\')', dir);
     this.court.judge(dir);
     return this.court.systems_suggestions;
   }