+++ src/generator/sugestion_chooser.js
@@ -1,4 +1,4 @@
-import { _ } from 'azk';
+import { _, log } from 'azk';
 import { UIProxy } from 'azk/cli/ui';
 
 var glob = require('glob');
@@ -26,13 +26,14 @@ export class SugestionChooser extends UIProxy {
   }
 
   suggest(evidences) {
-
     return _.map(evidences, (evidence) => {
       var suggestionChoosen = _.find(this.suggestions, (suggestion) => {
         var diff = _.difference([ evidence.ruleName ], suggestion.ruleNamesList);
         return diff.length === 0;
       });
-      evidence.suggestionChoosen = _.cloneDeep(suggestionChoosen);
+
+      evidence.suggestionChoosen            = _.clone(suggestionChoosen);
+      evidence.suggestionChoosen.suggestion = _.cloneDeep(suggestionChoosen.suggestion);
       return evidence;
     });
   }