+++ src/system/run.js
@@ -1,4 +1,4 @@
-import { _, t, Q, async, defer, config, lazy_require } from 'azk';
+import { _, t, Q, async, defer, config, lazy_require, log } from 'azk';
 import { ImageNotAvailable, SystemRunError, RunCommandError } from 'azk/utils/errors';
 import { Balancer } from 'azk/system/balancer';
 import net from 'azk/utils/net';
@@ -20,6 +20,7 @@ var Run = {
 
       if (_.isEmpty(steps)) return null;
       if ((!options.provision_force) && system.provisioned) return null;
+      log.debug('provision steps', steps);
 
       // provision command (require /bin/sh)
       var cmd  = ["/bin/sh", "-c", "( " + steps.join('; ') + " )"];
@@ -61,7 +62,6 @@ var Run = {
 
       yield this._check_image(system, options);
       var docker_opt = system.shellOptions(options);
-
       var container  = yield docker.run(system.image.name, command, docker_opt);
       var data       = yield container.inspect();