+++ src/utils/log.js
@@ -2,12 +2,14 @@ import { _, config, t } from 'azk';
 var winston = require('winston');
 
 var log = new winston.Logger();
+var sysUtil = require("util");
 
 // File log
 log.add(winston.transports.File, {
   filename: config('paths:log'),
   level: config('logs_level:file'),
   colorize: true,
+  prettyPrint: true,
   json: false,
 });
 
@@ -15,6 +17,7 @@ log.add(winston.transports.File, {
 var console_opts = {
   handleExceptions: true,
   colorize: true,
+  prettyPrint: true,
   level: config('logs_level:console')
 }
 if (config('env') != 'test') {
@@ -31,6 +34,6 @@ log.setConsoleLevel = (level) => {
   log.remove(winston.transports.Console);
   console_opts.level = level;
   log.add(winston.transports.Console, console_opts);
-}
+};
 
-export { log }
+export { log };